requirements.txt
Flask==3.0.0
pycryptodome==3.18.0
